ПРодолжение домашнего задания к 8 уроку.
