+++ server/package.json
@@ -14,5 +14,14 @@
     "cors": "^2.8.5",
     "express": "^4.18.1",
     "nodemon": "^2.0.16"
+  },
+  "devDependencies": {
+    "css-loader": "^6.7.1",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0"
   }
 }
