+++ server/public/basket_goods.json
@@ -1,6 +1,1 @@
-[
-    {
-        "id": 4,
-        "count": 1
-    }
-]
+[{"id":4,"count":1},{"id":3,"count":2},{"id":2,"count":1},{"id":1,"count":2}]
