+++ src/components/CustomButton/index.js
@@ -1,0 +1,7 @@
+export default Vue.component('custom-button', {
+    template: `
+    <button @click="$emit('click')" class="cart-button" type="button">
+      <slot></slot>
+    </button>
+    `
+})
