+++ src/index.js
@@ -1,0 +1,13 @@
+import './style.css';
+import './components/BasketItems';
+import { BASE_URL, GOODS_CART, GOODS, GOODS_BASKET } from './constants.js';
+import { service, serviceWithBody } from './services.js';
+import './components/Search';
+import './components/Basket';
+import './components/CustomButton';
+import './components/Good';
+import { app } from './app.js';
+
+window.onload = () => {
+
+}
